autonomous_agent_starter/requirements.txt
langchain==0.1.20
langchain-openai==0.1.8
langchain-community==0.0.38
fastapi==0.111.0
uvicorn[standard]==0.30.1
openai==1.35.3
redis==5.0.4
celery==5.3.4
chromadb==0.5.0
python-dotenv==1.0.1
psutil==5.9.8
requests==2.32.3
pydantic==2.7.4
